Crate http_handle

SSG Server

http-handle is a simple and efficient HTTP server for serving static websites generated by the Shokunin Static Site Generator (SSG). It handles HTTP requests and responses, manages server operations, and provides robust error handling.

Modules

server: Contains the core Server struct and logic for managing HTTP connections.

request: Handles incoming HTTP requests, parsing and validation.

response: Provides utilities for crafting HTTP responses.

error: Defines errors related to the server’s operation.

Re-exports

pub use error::ServerError;

pub use server::Server;

Modules

error

The error module defines various errors that can occur during server operation, including those related to connections and malformed requests. Error types for the Http Handle.

request

The request module is responsible for parsing and validating incoming HTTP requests. HTTP request parsing module for the Http Handle.

response

The response module provides tools and utilities for crafting HTTP responses.

server

The server module contains the core Server struct and associated methods for starting and managing the HTTP server.